requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
packaging==24.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
